Make changes to file structure and interfaces to prepare for conversion to merge-styles
