+++ packages/office-ui-fabric-react/src/components/Persona/Persona.types.ts
@@ -1,13 +1,15 @@
 import * as React from 'react';
 import { IRenderFunction } from '../../Utilities';
-import { Persona } from './Persona';
+import { PersonaBase } from './Persona.base';
 import { ImageLoadState } from '../../Image';
+import { IStyle, ITheme } from '../../Styling';
+import { IStyleFunction } from '../../Utilities';
 
 export interface IPersona {
 
 }
 
-export interface IPersonaProps extends React.HTMLAttributes<Persona> {
+export interface IPersonaProps extends React.HTMLAttributes<PersonaBase> {
   /**
    * Optional callback to access the IPersona interface. Use this instead of ref for accessing
    * the public methods and properties of the component.
@@ -139,6 +141,16 @@ export interface IPersonaProps extends React.HTMLAttributes<Persona> {
    * Optional HTML element props for Persona coin.
    */
   coinProps?: React.HTMLAttributes<HTMLDivElement>;
+
+  /**
+   * Call to provide customized styling that will layer on top of variant rules
+   */
+  getStyles?: IStyleFunction<IPersonaStyleProps, IPersonaStyles>;
+
+  /**
+   * Theme provided by HOC.
+   */
+  theme?: ITheme;
 }
 
 export enum PersonaSize {
@@ -223,4 +235,20 @@ export enum PersonaInitialsColor {
    * Its primary use is for overflow buttons, so it is considered a reserved color and can only be set with overrides.
    */
   transparent = 15,
+}
+
+export interface IPersonaStyleProps {
+  /**
+   * Theme.
+   */
+  theme: ITheme;
+
+  /**
+   * Custom class name.
+   */
+  className?: string;
+}
+
+export interface IPersonaStyles {
+  root: IStyle;
 }
