+++ packages/office-ui-fabric-react/src/components/Persona/index.ts
@@ -1,4 +1,5 @@
 export * from './Persona';
+export * from './Persona.base';
 export * from './Persona.types';
 export * from './PersonaConsts';
 export * from './PersonaCoin';
